ecs+resource panel in work

+++ Assets/_Scripts/Systems/TimeControlSystem.cs
@@ -10,11 +10,14 @@ public partial struct TimeControlSystem : ISystem
 {
     public void OnCreate(ref SystemState state)
     {
-        Entity _mainTimeState = state.EntityManager.CreateEntity(typeof(TimeState));
-        state.EntityManager.SetComponentData<TimeState>(_mainTimeState, new TimeState
+        if(!SystemAPI.HasSingleton<TimeState>())
         {
-            IsRunning = true
-        });
+            Entity mainTimeState = state.EntityManager.CreateEntity(typeof(TimeState));
+            state.EntityManager.SetComponentData<TimeState>(mainTimeState, new TimeState
+            {
+                IsRunning = true
+            });
+        }
     }
     public void OnDestroy(ref SystemState state)
     {
@@ -25,12 +28,14 @@ public partial struct TimeControlSystem : ISystem
     {
         if(!SystemAPI.HasSingleton<TimeToggleRequest>())
             { return; }
-        Entity toggleEntity = SystemAPI.GetSingletonEntity<TimeToggleRequest>();
 
-        TimeState timeState = SystemAPI.GetSingleton<TimeState>();
+        if(!SystemAPI.HasSingleton<TimeState>())
+            { return; }
+
+        ref TimeState timeState = ref SystemAPI.GetSingletonRW<TimeState>().ValueRW;
         timeState.IsRunning = !timeState.IsRunning;
-        SystemAPI.SetSingleton<TimeState>(timeState);
 
+        Entity toggleEntity = SystemAPI.GetSingletonEntity<TimeToggleRequest>();
         state.EntityManager.DestroyEntity(toggleEntity);
     }
 }
